🔨 refactor: 이미지 다운로드 유틸로 분리
