+++ src/pages/LetterStoragePage/ReplyStorage/ReplyPolaroidModal.tsx
@@ -1,37 +1,23 @@
-import axios from 'axios';
 import { css } from '@emotion/react';
 import PolaroidModal from '@/components/PolaroidModal';
 import Button from '@/components/Button';
 import IconButton from '@/components/IconButton';
 import { Download } from '@/assets/icons';
+import { imageDownload } from '@/utils/downloadUtils';
 
 interface ReplyPolaroidModalProps {
   imagePath: string;
 }
 
 const ReplyPolaroidModal = ({ imagePath }: ReplyPolaroidModalProps) => {
-  const handleDownload = async () => {
-    console.log('다운로드');
-    const response = await axios.get(imagePath, {
-      responseType: 'blob',
-    });
-
-    const url = window.URL.createObjectURL(new Blob([response.data]));
-
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = 'image.jpg';
-    document.body.appendChild(link);
-    link.click();
-
-    window.URL.revokeObjectURL(url);
-  };
-
   return (
     <PolaroidModal
       img={imagePath}
       headerRightContent={
-        <IconButton css={style.download} onClick={handleDownload}>
+        <IconButton
+          css={style.download}
+          onClick={() => imageDownload(imagePath)}
+        >
           <Download color="white" height={20} width={20} />
         </IconButton>
       }
